Create "tarea-destinos" exercise
Includes an HTML with initial elements, a js file with the first interactions and an angry cat gif that appears when the user choose not to play.

For now, the user is asked to play and to choose a "YES" or "NO" button. The NO button adds a "It's your loss!!!" message and a gif of an angry cat. The YES button asks another question about choosing a destination, with three buttons: France, Spain or Portugal (no functionality for now).